persist on event
