+++ crates/hyle-modules/src/node_state/module.rs
@@ -87,7 +87,7 @@ impl Module for NodeStateModule {
 
     async fn run(&mut self) -> Result<()> {
         module_handle_messages! {
-            on_bus self.bus,
+            on_self self,
             command_response<QueryBlockHeight, BlockHeight> _ => {
                 Ok(self.inner.current_height)
             }
