+++ src/p2p.rs
@@ -108,7 +108,7 @@ impl P2P {
         }
 
         module_handle_messages! {
-            on_bus self.bus,
+            on_self self,
             listen<NodeStateEvent> NodeStateEvent::NewBlock(b) => {
                 if b.block_height.0 > p2p_server.current_height {
                     p2p_server.current_height = b.block_height.0;
